Create common/application.js
Isolates functions to recreate or intereact with the application. It alleviates memory charge since we now load just the modules we need.
Currently there are just two application functions: init & cleanup. They are used just in Unit test suite but there are plans to use them in a new suite called System.