Avoid test index files
From now on, Javascript test files are automatically found by folder skipping those which contain the word common or are inside a common folder.
This mechanism avoid human-errors and assures to execute all the existing test files.

Some missing tests were discovered. The commit includes the easy-fixes for some of them or skip them including a TOFIX tag and linking to the specific open issue.