Fix protocol version parsing with double digits
The protocol version was being incorrectly parsed.
Using values with more than one digit on the hard
part caused the validation to only take in account
the first one.

Now the hard part and the soft part of the protocol
version are correctly parsed by being delimited by
a perdiod (.).